requirements.txt
# Kalshi News Pipeline — Python Dependencies
# Install: pip install -r requirements.txt

# ── Async HTTP ────────────────────────────────────────────────────────────────
httpx>=0.27.0           # async HTTP client (article fetching, GDELT, RSS)
websockets>=12.0        # Alpaca WebSocket connection

# ── RSS parsing ───────────────────────────────────────────────────────────────
feedparser>=6.0.11      # parse RSS/Atom feeds

# ── NLP — Fast (always required) ─────────────────────────────────────────────
vaderSentiment>=3.3.2   # lexicon-based sentiment (no model, ~1ms)

# ── NLP — Deep (optional but recommended) ────────────────────────────────────
# Comment out if you can't run a transformer model (no GPU / limited RAM)
transformers>=4.40.0    # HuggingFace pipeline (FinBERT)
torch>=2.0.0            # PyTorch backend for FinBERT
                        # → CPU-only install: pip install torch --index-url https://download.pytorch.org/whl/cpu

# ── Article extraction (optional but recommended) ────────────────────────────
newspaper3k>=0.2.8      # full article text extraction
lxml>=5.0.0             # HTML parsing backend for newspaper3k

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0    # load .env file for API keys (optional)

# ── Modal (remote GPU inference) ─────────────────────────────────────────────
modal>=0.73.0           # serverless GPU inference (set NLP_ARTICLE_MODE=modal)